libide/gtk: use GtkFilterListModel as basis for IdeUniqueListModel
This is a listmodel that removes subsequent items that are the same as
the previous item.

It still has soem faults, such as not handling insertions right when it
will match the exact next item (which is already shown). But we can deal
with that later as it doesn't affect us currently for the use of this
model in shortcuts.