plugins/meson: use /bin/sh -c for meson devenv
Instead of using brittle expansion bits manually, this sources a cached
meson devenv file when available and then executes the requested program
in a non-login/non-interactive shell.

We don't need the user shell for this because we're just executing tooling
and should be inheriting environment at other (deeper) layers of the
IdeRunContext stack.

Related #1945