plugins/flatpak: expose session a11y bus when running project
Instead of going through all the work to proxy an a11y bus, we can cheat
and just give the running application access to the session a11y bus. It
would be nice if this was closer to what Flatpak itself would do, but that
would require a lot of duplication-of-effort short of Flatpak exposing that
behavior as something we can reuse.

Fixes #1362