janitorial: port to GSignalGroup and GBindingGroup
This gets rid of our private copy and uses what is in GObject instead. To
reduce some churn in how we do borrowed references, a static inline is
used to keep the old behavior. Those can be fixed incrementally over time
without the need to keep the whole private copy of signal group and
binding group.