plugins/podman: always set CWD for podman-exec
The host runtime defaults to running in the user's home directory, so we
want to replicate that the best we can with the podman container. We do
incur some risk of breakage here in that perhaps there are people using
containers which do not have their home directory within it. If that
happens, we may need to choose something different. But for toolbox, the
main case we're covering here, that is to be expected.

Fixes #2042