libide-gui: unload workbenches on quit
Practically, this fixes failure to save the workbench session state when
quitting Builder with Ctrl+Q. When closing Builder with the mouse by
clicking the close button in the upper right (or left in RTL), Builder
remembers where you left off for next time you load the same project.
But when quitting with Ctrl+Q, it doesn't. Now it does. The main
difference is Ctrl+Q closes all windows instead of a single window, so
got to unload every workbench. This strategy was suggested by Christian.