plugins/flatpak: detect profiling with sysprof-agent and poke holes
We need to ensure we have access to a few places when running the app
because it's going to try to decode from within the Flatpak sandbox,
including things that were running on the host system.

It might be nice to defer symbol decoding until it's opened with the
Sysprof application, but that would ensure that the capture files are
not very share-able without extra work (ie: decode+attach symbols from
the UI application).

Related #2090