plugins/flatpak: clear invalid if network is not available
If we do not do this, we may accidentally trigger this to run after the
network portal *finally* tells us that the network is available. We don't
want that to happen because if it does, the flatpak-builder command which
runs to update the ostree cachepoint will have it's contents removed and
not run again after this.

All of this because the network portal is not giving correct information
on the first query.

Fixes #1939