plugins/symbol-tree: reduce timeout delay
Keyboard repeat seems to default to 30 msec for GNOME, so we just need
something that is larger than that to get value from the delay.

I hesitate to actually read the timeout value from GSettings because that
is the sort of thing that won't be available inside the sandbox without
going through the settings portal.

Related !757