libide/sourceview: use GtkTextBuffer::paste-done
This assumes we're in a paste until we get the GtkTextBuffer::paste-done
signal. I do worry slightly that there are situations where we might start
a paste and never get a paste-done. Currently that would mean that we just
don't continue inserting matching braces/overwrites/etc which is hardly
worse than the current situation.

If that ends up being a problem, we can create an issue for that and treat
it differently.

Fixes #1870