plugins/code-index: release resources after run
We keep the object-tree alive while processing, which probably should be
addressed at some point. (It does clean up at the end). But this helps
drop a bunch of resources as soon as they are no longer necessary.

With this and the previous patch I can keep the kernel to around 1Gb
resident while building the search indexes, and have it drop back down to
where it should be afterwards.

This likely fixes the majority of #2166