libide/core: introduce global cache root with project overrides
This introduces a new GSetting for a "cache root" other than the previous
.cache/gnome-builder/ (or XDG equivalent). You may set the default for all
projects (/home/user/Projects/Builds for example) and it also allows for
per-project overrides should that be necessary.

This may help in situations where you want to want to ensure your build
artifacts are on faster/slower/unmirrored/etc storage than your home
directory.

We may want to followup on this because there are a couple things which
could still be addressed further.

 * Flatpak SDKs could be moved here (moving our entire installation
   into this directory). It would be nice if we could actually move the
   installation into place, but I'm not sure if that will actually work
   once we have composefs type things for OSTree.

 * We may want to choose a default for this such as
   ~/Projects/.gnome-builder or something like that.

Fixes: #1600