plugins/manuals: add pathbar to statusbar
When viewing documentation, show the path to the item in the pathbar.

This gets the viewing in place, but there is still more work to be done so
that we can do the same hooks the full app did. For example, we need to
hijack the browsing action still. We also need to allow revealing the
navigatable in the sidebar so all element types work correctly.