plugins/manuals: setup decide-policy helpers
This will redirect links to things we are pretty certain already exist
within our repository to the native object/navigatable instead of jumping
to the web for them.

If it doesn't exist in our repository, we jump to the web anyway to
signify to the user this isn't documentation we control.