plugins/manuals: implement search sections for queries
This makes ManualsSearchQuery a GtkSectionModel so that we can group
results by their respective SDKs without having to load every single
item into memory.

This sort of cheats to do that by doing a query for each SDK. But in
practice that's a small number anyway so should be fine.