libide/webkit: work around webkit DMA-BUF issue
Previous to this (with new enough WebKit supporting DMA-BUF) we would see
an error about not being able to create a DMA-BUF on the console.

 Failed to create GBM buffer of size 0x0: Invalid argument

Ultimately, to work around this we need to create our window, present it
so that it gets mapped and resources created, then wait a little bit of
time, then add the page w/ WebKit (so it can successfully create it's
resources), then load the URI.

Otherwise, we get a window with proper window titles/etc but no contents
rendered to the page.

Fixes #2186