libide/sourceview: remove dead accelerators for change-case
These aren't actually mapped anywhere so we should remove them from the
list of accels that users might see in the shortcuts dialog.

It would be nice to follow up on this and allow users to map them to
something specifically, but this at least removes any sort of perceived
accel clash.

Fixes #2211