libide/tree: simplify tree using libdex
This moves the building code to using a fiber which is a bit easier to
think about and maintain than the whole chain of function state we had
before.

Additionally, we try to reduce the IdeExtensionSetAdapter to just the
tree and use GListModel after that. This will allow for alternate ways
of populating builders.