plugins/flatpak: override --libdir= for various build systems
This tries a quick workaround for --libdir to help ensure that we are
configuring things the same as flatpak-builder would.

This isn't really ideal in that we will end up showing it to the user
in the configuration dialog. But since we cannot handle this from the
build systems directly (as it's a huge layering violation), we have to
modify it somewhere.

This will not write-back though so it only becomes persistent if the
user modifies the configuration line manually afterwards.

Fixes #2227