shellcmd: Don't treat command & name as markup
Previously, in the list of commands, the command to be executed and the
name it has been given would both be treated as Pango markup. This
meant that if you named a command “Build & Run”, or defined a command
like “sh -c 'foo && bar'”, these would not be displayed correctly in the
list.

Set AdwActionRow:use-markup to FALSE to fix this.