libide/io: make recursive file monitor ignore check async
This allows asychronously querying for ignored status from the recursive
file monitor instead of forcing it to be synchronous. Doing so helps avoid
some UI lockups at startup even though we have to do _more_ work in this
mode.

There are other places still doing synchronous ignore checking, but we
will see if that is an issue after this is integrated.

Related: 2235