Add European Union Public Licence templates
The EUPL-1.2 is as a licence written to be compatible with recent
versions of the GPL among other licences but written in such a way that
it is better adapted to the copyright law of European Union member
states. There are two previous versions, 1.1 and 1.0, but I am not sure
whether there is any particular value in including those given that they
are older versions and I am not sure whether they have any reasons to be
preferred.

Additionally, according to SPDX, there is no standard licence header for
the EPL-1.2[1]. As such, I wrote one based on the MPL-2.0 licence header
already used in Builder.

 [1]: https://spdx.org/licenses/EUPL-1.2.html