libide/gui: improve re-use of tweaks within workbench
We still allow creating more than one of these so they can be attached to
the workbench's window group (and also show up on the appropriate desktop
workspace when multiple are used), but this ensures we re-use them in the
common case.

It also tries to handle the greeter/editor workspace a bit better.