lib/tree: fix delayed-determination of children-possible
GtkTreeListModel cannot be updated when children-possible changes. That
means for GtkListView to get the necessary updates we must notify that
the parent model (e.g. parent IdeTreeNode) has removed/inserted a new
row. It is the same row, of course but now can be re-built and have the
children discovered.

Fixes: #2321