buffer: apply diagnostics from fiber in 1msec intervals
There are some side-effects to this, such as we might get the wrong line
for diagnostics if the user keeps editing. But that exists already given
the latency between doing the diagnostics call and getting the results.

We will just improve the location on the next application of diagnostics.

Related: #2346